fix issue 76
