+++ models/tts/valle/valle_dataset.py
@@ -289,24 +289,3 @@ def batch_by_size(
     if len(batch) > 0:
         batches.append(batch)
     return batches
-
-
-class VariableSampler(BatchSampler):
-    def __init__(self, sampler, drop_last: bool, use_random_sampler=False):
-        self.data_list = sampler
-        if use_random_sampler:
-            self.sampler = RandomSampler(sampler)
-        else:
-            self.sampler = SequentialSampler(sampler)
-
-        super().__init__(self.sampler, 1, drop_last)
-
-    def __iter__(self):
-        for batch_ids in self.data_list:
-            yield batch_ids
-
-    def __len__(self):
-        if self.drop_last:
-            return len(self.sampler) // self.batch_size
-        else:
-            return (len(self.sampler) + self.batch_size - 1) // self.batch_size
